+++ package.json
@@ -21,5 +21,8 @@
     "eslint-plugin-jsx-a11y": "^6.2.1",
     "eslint-plugin-react": "^7.13.0",
     "prop-types": "^15.7.2"
+  },
+  "dependencies": {
+    "react-router-dom": "^5.0.1"
   }
 }
