+++ todo/src/components/About/About.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const About = () => (
+  <div>
+    About
+  </div>
+);
+
+export default About;
