+++ todo/src/components/App/App.module.css
@@ -1,9 +1,23 @@
-.todo {
+.App {
+  display: flex;
   margin: auto;
-  width: 300px;
+  width: 400px;
   font-family: 'Montserrat', sans-serif;
+}
+
+.content {
+  width: 300px;
 }
 
 .title {
   font-weight: bold;
+}
+
+.sidebar {
+  width: 100px;
+}
+
+.link {
+  color: rgba(0, 0, 0, 0.87);
+  text-decoration: none;
 }
