+++ todo/src/components/Contacts/Contacts.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Contacts = () => (
+  <div>
+    About
+  </div>
+);
+
+export default Contacts;
